+++ src/app/login/index.js
@@ -1,11 +1,11 @@
 import React, {useCallback} from "react";
 import useStore from "../../hooks/use-store";
-import LayoutFlex from "../../components/layout-flex";
-import Layout from "../../components/layout";
+import LayoutFlex from "../../components/layouts/layout-flex";
+import Layout from "../../components/layouts/layout";
 import LocaleSelect from "../../containers/locale-select";
 import useTranslate from "../../hooks/use-translate";
 import Tools from "../../containers/tools";
-import LoginForm from "../../components/login-form";
+import LoginForm from "../../components/form-components/login-form";
 import useSelector from "../../hooks/use-selector";
 
 function Login() {
