+++ src/app/main/index.js
@@ -5,8 +5,8 @@ import useTranslate from "../../hooks/use-translate";
 import CatalogFilter from "../../containers/catalog-filter";
 import CatalogList from "../../containers/catalog-list";
 import Tools from "../../containers/tools";
-import LayoutFlex from "../../components/layout-flex";
-import Layout from "../../components/layout";
+import LayoutFlex from "../../components/layouts/layout-flex";
+import Layout from "../../components/layouts/layout";
 import LocaleSelect from "../../containers/locale-select";
 
 function Main() {
