+++ src/app/profile/index.js
@@ -1,7 +1,6 @@
 import React from "react";
-import useStore from "../../hooks/use-store";
-import LayoutFlex from "../../components/layout-flex";
-import Layout from "../../components/layout";
+import LayoutFlex from "../../components/layouts/layout-flex";
+import Layout from "../../components/layouts/layout";
 import LocaleSelect from "../../containers/locale-select";
 import useTranslate from "../../hooks/use-translate";
 import Tools from "../../containers/tools";
@@ -9,15 +8,13 @@ import ProfileCard from "../../components/profile-card";
 import useSelector from "../../hooks/use-selector";
 
 function Profile(props) {
-  const store = useStore();
   
   const {t} = useTranslate();
-  const {userName, userEmail, userPhone, token} = useSelector(state => {
+  const {userName, userEmail, userPhone} = useSelector(state => {
     return {
       userName: state.user.user?.profile?.name,
       userEmail: state.user.user?.email,
       userPhone: state.user.user?.profile?.phone,
-      token: state.user.token,
     }
   });
