+++ src/components/basket-components/item-basket/index.js
@@ -1,6 +1,6 @@
 import React, {useCallback} from 'react';
 import propTypes from 'prop-types';
-import numberFormat from "../../utils/number-format";
+import numberFormat from "../../../utils/number-format";
 import {cn as bem} from "@bem-react/classname";
 import {Link} from "react-router-dom";
 import './styles.css';
