--- src/components/basket-simple/styles.css
@@ -1,16 +1,0 @@
-/* Контрол с общими действиями */
-.BasketSimple{
-    display: flex;
-    justify-content: flex-end;
-    padding: 0;
-    font-size: 18px;
-}
-
-.BasketSimple-label{
-
-}
-
-.BasketSimple-total{
-    font-weight: bold;
-    margin: 0 15px;
-}
