--- src/components/basket-total/styles.css
@@ -1,16 +1,0 @@
-/* Пункт списка */
-.BasketTotal {
-    display: flex;
-    justify-content: flex-end;
-    padding: 20px;
-}
-
-.BasketTotal-cell {
-    min-width: 85px;
-    text-align: right;
-    font-weight: bold;
-    font-size: 18px;
-    padding-left: 10px;
-    white-space: nowrap;
-}
-
