+++ src/components/form-components/input/style.css
@@ -1,0 +1,8 @@
+.Input{
+    font-size: 16px;
+    padding: 2px;
+}
+
+.Input_theme_big{
+    width: 300px;
+}
