+++ src/components/form-components/login-form/style.css
@@ -1,7 +1,6 @@
-/* форма для авторизации */
+/* Форма для авторизации */
 .LoginForm{
 margin: 0 20px;
-
 }
 
 .LoginForm-title {
