+++ src/components/form-components/select/style.css
@@ -1,0 +1,4 @@
+.Select{
+    font-size: 16px;
+    padding: 3px 2px 2px;
+}
