--- src/components/input/style.css
@@ -1,8 +1,0 @@
-.Input{
-    font-size: 16px;
-    padding: 2px;
-}
-
-.Input_theme_big{
-    width: 300px;
-}
