--- src/components/item-basket/styles.css
@@ -1,40 +1,0 @@
-/* Пункт списка */
-.ItemBasket {
-    display: flex;
-    justify-content: flex-start;
-    transition: background-color 1s;
-    padding-left: 10px;
-}
-
-.ItemBasket-id {
-    padding: 20px 10px 20px 20px;
-}
-
-.ItemBasket-title {
-    width: 100%;
-    padding: 20px 10px;
-}
-
-.ItemBasket-title a{
-    color: #000;
-    text-decoration: none;
-}
-
-.ItemBasket-title a:hover{
-    color: #000;
-    text-decoration: underline;
-}
-
-.ItemBasket-right {
-    padding: 20px 20px 20px 10px;
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-}
-
-.ItemBasket-cell {
-    min-width: 95px;
-    text-align: right;
-    font-size: 18px;
-    white-space: nowrap;
-}
