--- src/components/layout/style.css
@@ -1,24 +1,0 @@
-body{
-    margin: 0;
-    padding: 0;
-    background: #535353;
-    font-family: sans-serif;
-    position: relative;
-}
-/* Страница*/
-.Layout{
-    max-width: 1024px;
-    min-height: calc(100vh - 42px);
-    margin: 0 auto;
-    background: #fff;
-    box-sizing: border-box;
-    /*box-shadow: 0 0 15px 5px #0000003b;*/
-}
-.Layout-head{
-    height: 80px;
-    background: #f5f5f5;
-    padding: 0;
-}
-.Layout-head h1{
-    margin: 0;
-}
