--- src/components/select/style.css
@@ -1,4 +1,0 @@
-.Select{
-    font-size: 16px;
-    padding: 3px 2px 2px;
-}
