+++ src/containers/tools/index.js
@@ -3,8 +3,8 @@ import useSelector from "../../hooks/use-selector";
 import useStore from "../../hooks/use-store";
 import useTranslate from "../../hooks/use-translate";
 import Menu from "../../components/menu";
-import BasketSimple from "../../components/basket-simple";
-import LayoutFlex from "../../components/layout-flex";
+import BasketSimple from "../../components/basket-components/basket-simple";
+import LayoutFlex from "../../components/layouts/layout-flex";
 
 function Tools() {
